pyproject.toml
[project]
name = "fastapi-demo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.1",
    "langchain>=1.0.5",
    "langchain-deepseek>=1.0.0",
    "langgraph-checkpoint-postgres>=3.0.1",
    "minio>=7.2.18",
    "orjson>=3.11.4",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "qiniu>=7.17.0",
    "redis>=7.1.0",
    "selenium>=4.38.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
]
